move parsePost, change err check
